// src/app/_dto/irepparam.dto.ts
export class IrepparamDto {

    menuId: string;

    optionName: string;

    optionLabel: string | null;

    optionValue: string | null;

    defaultValue: string;

    optionTitle: string | null;

    visibility: string = 'true';

    optionType: string | null;

    criteria: string | null;

    optionId: string | null;

    reportFile: string | null;

}